requirements.txt
langchain-google-genai>=0.0.3
python-dotenv>=1.0.0
google-generativeai>=0.3.0
fastapi>=0.104.1
uvicorn>=0.24.0
beautifulsoup4>=4.12.2
requests>=2.31.0
